refactor: 한 번만 export하는 변수들을 각각의 파일로 이동
