+++ hooks/src/lib/constants/cardConfig.ts
@@ -1,21 +1,8 @@
-export const CARD_NUMBER = {
-  maxLength: 4,
-  fieldCount: 4,
-};
-
 export const CARD_EXPIRATION = {
   minMonth: 1,
   maxMonth: 12,
   minYear: 25,
   monthLength: 2,
   yearLength: 2,
-  maxLength: 2,
-};
-
-export const CARD_CVC = {
-  maxLength: 3,
-};
-
-export const CARD_PASSWORD = {
   maxLength: 2,
 };
